perf: lazy-load the chatbot implementation on visitor intent
The full chat drawer (react-markdown rendering, lead-capture form,
useLeadCapture, Gemini calls) shipped in the main bundle even for
visitors who never open it. AIChat.tsx now stays a cheap, always-mounted
trigger button plus intent detection (click, keyboard, the
`toggle-ai-chat` custom event, and `?chat=open`/`?chat=1` deep links);
the heavy implementation moved to AIChatPanel.tsx, loaded via
React.lazy only once one of those signals fires, wrapped in the
existing ChunkErrorBoundary + a Suspense fallback with an accessible
loading state.

main-entry gzip: 156.8 KB -> 86.7 KB (-45%); AIChatPanel lazy chunk:
46.4 KB gzip, budgeted in bundlesize.config.json.

Closes #1140.
